Rename UntiledMode to RegionRenderer; drop the ZarrMode abstraction
