+++ src/renderer-types.ts
@@ -1,3 +1,6 @@
+import type { MercatorBounds, Wgs84Bounds } from './map-utils'
+import type { ProjectionData, ShaderData } from './shaders'
+
 export interface RendererUniforms {
   clim: [number, number]
   opacity: number
@@ -42,3 +45,50 @@ export type ProjectionMode =
   | 'mapbox'
   | 'mapbox-proj4'
   | 'mapbox-ecef'
+
+export interface RenderContext {
+  gl: WebGL2RenderingContext
+  matrix: number[] | Float32Array | Float64Array
+  uniforms: RendererUniforms
+  colormapTexture: WebGLTexture
+  worldOffsets: number[]
+  customShaderConfig?: CustomShaderConfig
+  shaderData?: ShaderData
+  projectionData?: ProjectionData
+  mapbox?: MapboxParams
+  isGlobe?: boolean
+}
+
+/** Identifies a Mapbox tile for the draped renderToTile path. */
+export interface TileId {
+  z: number
+  x: number
+  y: number
+}
+
+export interface RegionRenderState {
+  texture: WebGLTexture
+  vertexBuffer: WebGLBuffer
+  /** Texture coordinate buffer for sampling resampled data */
+  pixCoordBuffer: WebGLBuffer
+  vertexArr: Float32Array
+  mercatorBounds: MercatorBounds
+  width: number
+  height: number
+  channels: number
+  /** Data orientation: true = row 0 is south */
+  latIsAscending: boolean
+  /** Band textures for multi-band custom shaders */
+  bandData?: Map<string, Float32Array>
+  bandTextures?: Map<string, WebGLTexture>
+  bandTexturesUploaded?: Set<string>
+  bandTexturesConfigured?: Set<string>
+  /** Index buffer for adaptive source-projected meshes */
+  indexBuffer?: WebGLBuffer
+  /** Number of vertices/indices to draw */
+  vertexCount?: number
+  /** Whether to use indexed mesh rendering (gl.drawElements) */
+  useIndexedMesh?: boolean
+  /** WGS84 bounds for two-stage reprojection */
+  wgs84Bounds?: Wgs84Bounds
+}
